Added MISU

+++ src/train.py
@@ -305,7 +305,7 @@ def step(
             if type(segmentation) == list:
                 loss = torch.sum(torch.stack([criterion(s, ground_truth) for s in segmentation]))
             else:
-                loss = criterion(segmentation, ground_truth)
+                loss = criterion(segmentation, ground_truth.float())
             patients_perf.append(dict(id=patient_id[0], epoch=epoch, split=mode, loss=loss.item()))
 
             # Checking not nan value
